Add new todo service and dependencies
This commit adds a new todo service and its corresponding dependencies to the project. The service is responsible for handling the insertion of new todo items into the database. It uses the SQLiteWrapper class to interact with the SQLite database. The service takes a NewTodoPayload object as input, which contains the title, description, and user_id of the new todo item. The service builds the database connection and inserts the new todo item into the "todos" table.